feat(iam): SCRUM-51 Integrate Monolog for logging and error tracking
Logs are necessary to track occurred errors, because they may have
sensitive information, that a client do not permitted to see. Therefore,
choose Monolog as it is the most popular(*) PHP logging library, which
Symfony integrates seamlessly with. It fully covers both the project's
requirements and future use cases.

(*) According to the official Symfony documentation:
https://symfony.com/doc/current/logging.html#monolog